Fix memory leak in batiment.cc: turned pointer into a local member
There was no destructor for the Batiment class. Also, Batiment and Tile
shared a pointer to the same Coordinate object when a Tile was build
using a Batiment.

I didn't find any use for this sharing of data (for instance changing
the coordinates of a Batiment and see the change in Tile) so I turned
the pointers into local members.

Note that Batiment and Tiles now begin with a Coordinate(0,0) object
instead of an uninitialized pointer as their `picture` member.