repairing documentation + checking c++11
And there it is, the most waited commit on earth : You do not have to
make the doc in order to be able to compile.
the new command to make the doc is : make doxygen-doc
Feel free to change doxigen.cfg (config file) and to add css sheet.
Please do not push the doc on git.